Multiple changes for refactoring....

+++ ffead-server-freebsd/include/Timer.h
@@ -33,11 +33,11 @@ public:
 	virtual ~Timer();
 	void start();
 	void end();
-	static long getCurrentTime();
-	static long getTimestamp();
-	int elapsedMicroSeconds();
-	int elapsedMilliSeconds();
-	int elapsedNanoSeconds();
+	static long long getCurrentTime();
+	static long long getTimestamp();
+	long elapsedMicroSeconds();
+	long elapsedMilliSeconds();
+	long elapsedNanoSeconds();
 	int elapsedSeconds();
 };
